[DESIGN] 일지 상세 페이지 - 일상 UI 완성
