+++ app/src/main/java/com/kittyandpuppy/withallmyanimal/detail/DetailBehaviorActivity.kt
@@ -1,12 +1,12 @@
-package com.kittyandpuppy.withallmyanimal
+package com.kittyandpuppy.withallmyanimal.detail
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.withallmyanimal.R
 
-class DetailHospitalActivity : AppCompatActivity() {
+class DetailBehaviorActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_detail_hospital)
+        setContentView(R.layout.activity_detail_behavior)
     }
 }
