+++ app/src/main/res/layout/activity_detail_hospital.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.kittyandpuppy.withallmyanimal.DetailHospitalActivity">
+    tools:context="com.kittyandpuppy.withallmyanimal.detail.DetailHospitalActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/tb_detail_hospital"
