+++ app/src/main/res/layout/activity_detail_pet.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.kittyandpuppy.withallmyanimal.DetailPetActivity">
+    tools:context="com.kittyandpuppy.withallmyanimal.detail.DetailPetActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/tb_detail_pet"
